+++ tests/Spec/GHC/TypeLitsSpec.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 
 {-|
@@ -22,6 +23,9 @@ import Prelude ()
 import Prelude.Compat
 
 import Spec.Utils (matchesTextShowSpec)
+#if MIN_VERSION_base(4,18,0)
+import Spec.Utils (Some)
+#endif
 
 import Test.Hspec (Spec, describe, hspec, parallel)
 
@@ -34,3 +38,15 @@ spec = parallel $ do
         matchesTextShowSpec (Proxy :: Proxy SomeNat)
     describe "SomeSymbol" $
         matchesTextShowSpec (Proxy :: Proxy SomeSymbol)
+#if MIN_VERSION_base(4,16,0)
+    describe "SomeChar" $
+        matchesTextShowSpec (Proxy :: Proxy SomeChar)
+#endif
+#if MIN_VERSION_base(4,18,0)
+    describe "Some SNat" $
+        matchesTextShowSpec (Proxy :: Proxy (Some SNat))
+    describe "Some SSymbol" $
+        matchesTextShowSpec (Proxy :: Proxy (Some SSymbol))
+    describe "Some SChar" $
+        matchesTextShowSpec (Proxy :: Proxy (Some SChar))
+#endif